Update Interaction
Add Interaction Wait (Execute, Trigger)
