+++ Assets/Editor/JsonUtility/JsonUtility/Json/Stage1/tttInteract.json.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: aac1d331015bc4e1bb5569134f43c185
+TextScriptImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
