+++ Assets/Scripts/Utility/Dialogue.meta
@@ -1,0 +1,3 @@
+fileFormatVersion: 2
+guid: 033031c5e6b442cbb1c6703546e4c9da
+timeCreated: 1663126058
