+++ Assets/Scripts/Utility/Dialogue/DialogueData.cs.meta
@@ -1,0 +1,3 @@
+fileFormatVersion: 2
+guid: f3113dc112e54fa98cbef998f60a9e7e
+timeCreated: 1663127503
