+++ Assets/Scripts/Utility/Dialogue/Enum.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: 74dc238b5f1449488072134b1207ec2b
+timeCreated: 1662567696
