+++ Assets/Scripts/Utility/Interaction/InputInteraction.cs.meta
@@ -1,0 +1,3 @@
+fileFormatVersion: 2
+guid: 901e0f19339a42559e093b794488befe
+timeCreated: 1678082268
