+++ Assets/Scripts/Utility/Interaction/Interaction.cs.meta
@@ -1,0 +1,3 @@
+fileFormatVersion: 2
+guid: 2a9c90ab00014bdbb690d8ffc6730ed8
+timeCreated: 1678436124
