+++ Assets/Scripts/Utility/Interaction/InteractionData.cs
@@ -1,0 +1,21 @@
+using System;
+using UnityEngine;
+
+namespace Utility.Interaction
+{
+    [Serializable]
+    public class InteractionData
+    {
+        public TextAsset jsonAsset;
+
+        public bool isInteractable;
+
+        public bool isContinuable;
+        public bool useNextInteract;
+
+        [Header("For Debugging")] public bool isInteracted;
+
+        public Action onInteractionStart;
+        public Action onInteractionEnd;
+    }
+}
