+++ Assets/Scripts/Utility/Interaction/InteractionData.cs.meta
@@ -1,0 +1,3 @@
+fileFormatVersion: 2
+guid: a7ee10e17f774850a2b87f3bab74f760
+timeCreated: 1678445244
