+++ Assets/Scripts/Utility/Interaction/TriggerInteraction.cs.meta
@@ -1,0 +1,3 @@
+fileFormatVersion: 2
+guid: 31d7567dfc1d44cc84f162ee42134ac8
+timeCreated: 1678440253
