--- Assets/Scripts/Utility/UI/Dialogue.meta
@@ -1,3 +1,0 @@
-fileFormatVersion: 2
-guid: 033031c5e6b442cbb1c6703546e4c9da
-timeCreated: 1663126058
