--- Assets/Scripts/Utility/UI/Dialogue/DialogueController.cs.meta
@@ -1,3 +1,0 @@
-fileFormatVersion: 2
-guid: 8a0f91c685e440538fed2176fb59327d
-timeCreated: 1663126091
