+++ ProjectSettings/TagManager.asset
@@ -19,7 +19,7 @@ TagManager:
   - NPC Interactor
   - Always Light
   - Light
-  - 
+  - OnlyPlayerCheck
   - 
   - 
   -
